NumberSample - Split file logic
