+++ NumberSample/NumberSample/CurrencyTextField.swift
@@ -1,0 +1,28 @@
+//
+//  CurrencyTextField.swift
+//  NumberSample
+//
+//  Created by Benoit Pasquier on 10/2/22.
+//
+
+import Foundation
+import SwiftUI
+
+struct CurrencyTextField: UIViewRepresentable {
+
+    typealias UIViewType = CurrencyUITextField
+
+    let numberFormatter: NumberFormatterProtocol
+    let currencyField: CurrencyUITextField
+
+    init(numberFormatter: NumberFormatterProtocol, value: Binding<Int>) {
+        self.numberFormatter = numberFormatter
+        currencyField = CurrencyUITextField(formatter: numberFormatter, value: value)
+    }
+
+    func makeUIView(context: Context) -> CurrencyUITextField {
+        return currencyField
+    }
+
+    func updateUIView(_ uiView: CurrencyUITextField, context: Context) { }
+}
